Jul 08: Bug Fixed 'advanced_loan_management'

+++ advanced_loan_management/models/loan_type.py
@@ -30,7 +30,7 @@ class LoanType(models.Model):
     _description = 'Loan Type'
 
     name = fields.Char(string='Name', help="LoanType Name", required=True)
-    loan_amount = fields.Integer(string='Loan Amount', help="Loan Amount", required=True)
+    loan_amount = fields.Float(string='Loan Amount', help="Loan Amount", required=True)
     tenure = fields.Integer(string='Tenure', default='1',
                             help="Amortization period")
     tenure_plan = fields.Char(string="Tenure Plan", default='Monthly',
